refactor: remove wishtack, replace lazy components with generated
