+++ package.json
@@ -63,7 +63,6 @@
     "@ngx-translate/http-loader": "^4.0.0",
     "@sentry/browser": "^5.15.5",
     "@trademe/ng-defer-load": "^8.0.0",
-    "@wishtack/reactive-component-loader": "^1.1.2",
     "angular2-cookie-law": "^7.0.1",
     "angular2-uuid": "^1.1.1",
     "angulartics2": "^9.0.0",
@@ -75,7 +74,6 @@
     "express-robots-txt": "0.4.1",
     "lodash-es": "^4.17.15",
     "morgan": "^1.10.0",
-    "ng-dynamic-component": "^6.1.0",
     "ng-recaptcha": "^5.0.0",
     "ng2-validation": "^4.2.0",
     "ngx-infinite-scroll": "^8.0.2",
