+++ src/app/extensions/quickorder/exports/.gitignore
@@ -1,0 +1,1 @@
+*/**/*.*
