--- src/app/extensions/quickorder/exports/quickorder/lazy-header-quickorder/lazy-header-quickorder.component.html
@@ -1,1 +1,0 @@
-<wt-lazy *ishFeature="'quickorder'" [location]="componentLocation"></wt-lazy>
