--- src/app/extensions/quickorder/exports/quickorder/lazy-header-quickorder/lazy-header-quickorder.component.ts
@@ -1,15 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-
-@Component({
-  selector: 'ish-lazy-header-quickorder',
-  templateUrl: './lazy-header-quickorder.component.html',
-  changeDetection: ChangeDetectionStrategy.Default,
-})
-
-// tslint:disable-next-line:component-creation-test
-export class LazyHeaderQuickorderComponent {
-  componentLocation = {
-    moduleId: 'ish-extensions-quickorder',
-    selector: 'ish-header-quickorder',
-  };
-}
