+++ src/app/extensions/quoting/exports/.gitignore
@@ -1,0 +1,1 @@
+*/**/*.*
