--- src/app/extensions/quoting/exports/account/lazy-quote-widget/lazy-quote-widget.component.html
@@ -1,1 +1,0 @@
-<wt-lazy *ishFeature="'quoting'" [location]="componentLocation"></wt-lazy>
