--- src/app/extensions/quoting/exports/basket/lazy-basket-add-to-quote/lazy-basket-add-to-quote.component.html
@@ -1,1 +1,0 @@
-<wt-lazy *ishFeature="'quoting'" [location]="componentLocation"></wt-lazy>
