--- src/app/extensions/quoting/exports/product/lazy-product-add-to-quote/lazy-product-add-to-quote.component.html
@@ -1,5 +1,0 @@
-<wt-lazy
-  *ishFeature="'quoting'"
-  [location]="componentLocation"
-  [inputs]="{ product: product, disabled: disabled, displayType: displayType, class: class, quantity: quantity }"
-></wt-lazy>
