+++ src/app/extensions/wishlists/exports/.gitignore
@@ -1,0 +1,1 @@
+*/**/*.*
