--- src/app/extensions/wishlists/exports/account/lazy-wishlist-widget/lazy-wishlist-widget.component.html
@@ -1,1 +1,0 @@
-<wt-lazy *ishFeature="'wishlists'" [location]="componentLocation"></wt-lazy>
