--- src/app/extensions/wishlists/exports/products/lazy-product-add-to-wishlist/lazy-product-add-to-wishlist.component.html
@@ -1,5 +1,0 @@
-<wt-lazy
-  *ishFeature="'wishlists'"
-  [location]="componentLocation"
-  [inputs]="{ product: product, displayType: displayType }"
-></wt-lazy>
