--- src/app/extensions/wishlists/exports/wishlists/lazy-wishlist-link/lazy-wishlist-link.component.html
@@ -1,1 +1,0 @@
-<wt-lazy *ishFeature="'wishlists'" [location]="componentLocation" [inputs]="{ view: view }"></wt-lazy>
