+++ src/app/shell/header/header-default/header-default.component.html
@@ -20,7 +20,7 @@
       <ish-product-compare-status data-testing-id="compare-status-desktop"> </ish-product-compare-status>
     </li>
     <li *ishFeature="'wishlists'" class="d-none d-md-block">
-      <ish-lazy-wishlist-link></ish-lazy-wishlist-link>
+      <ish-lazy-wishlists-link></ish-lazy-wishlists-link>
     </li>
   </ul>
