+++ src/app/shell/header/user-information-mobile/user-information-mobile.component.html
@@ -5,7 +5,7 @@
     <ish-product-compare-status view="full"></ish-product-compare-status>
   </div>
   <div *ishFeature="'wishlists'" class="user-info-box-item">
-    <ish-lazy-wishlist-link view="full"></ish-lazy-wishlist-link>
+    <ish-lazy-wishlists-link view="full"></ish-lazy-wishlists-link>
   </div>
   <div class="user-info-box-item"><ish-language-switch view="accordion" placement="up"></ish-language-switch></div>
 </div>
